+++ src/checkboxes/checkbox.cc
@@ -182,7 +182,7 @@ bool CDetectCheckBoxes::detectCheckBoxes(Mat &imgMATgray, vector<CCheckBox> &cBo
     vector<bool> usedInner(holes.size(), false);
 
 
-    for(int i=0; i<holes.size(); i++){
+    for(size_t i=0; i<holes.size(); i++){
 
         if(isAlmostSquare(holes[i])){
             CCheckBox cb;
@@ -201,7 +201,7 @@ bool CDetectCheckBoxes::detectCheckBoxes(Mat &imgMATgray, vector<CCheckBox> &cBo
 
     fillRatioFilter(imgMATbin, cBoxes);
 
-    for(int i=0; i<cBoxes.size(); i++){
+    for(size_t i=0; i<cBoxes.size(); i++){
         CCheckBox cb = cBoxes[i];
         if(cb.innerFillRatio < 0.05){
             cBoxes[i].isFilled = false;
@@ -270,7 +270,7 @@ bool CDetectCheckBoxes::detectCheckBoxes(string &inFileName, vector<CCheckBox> &
         height.push_back(atoi(temp.c_str()));
     }
 
-    for (int j = 0; j < x.size(); j++)
+    for (size_t j = 0; j < x.size(); j++)
         fprintf(stderr, "%d " "%d " "%d " "%d \n", x[j], y[j], width[j], height[j]);
 
     vector<Rect> holes;
@@ -281,7 +281,7 @@ bool CDetectCheckBoxes::detectCheckBoxes(string &inFileName, vector<CCheckBox> &
         waitKey(0);
         destroyWindow("out");
         Preprocessor::conCompFast(new_mat, holes, 0.9, 0.9, 200, 4);
-        for (int k = ex_size; k < holes.size(); k++){
+        for (size_t k = ex_size; k < holes.size(); k++){
             holes[k].x += x[i];
             holes[k].y += y[i];
         }
@@ -299,7 +299,7 @@ bool CDetectCheckBoxes::detectCheckBoxes(string &inFileName, vector<CCheckBox> &
         waitKey(0);
         destroyWindow("out");*/
         Preprocessor::conCompFast(new_mat, rboxes, 0.9, 0.9, 200, 4);
-        for (int k = ex_size; k < rboxes.size(); k++){
+        for (size_t k = ex_size; k < rboxes.size(); k++){
             rboxes[k].x += x[i];
             rboxes[k].y += y[i];
         }
@@ -308,12 +308,12 @@ bool CDetectCheckBoxes::detectCheckBoxes(string &inFileName, vector<CCheckBox> &
     fprintf(stderr, "%s\n", "done!");
 
     vector<bool> usedInner(holes.size(), false);
-    for(int i=0; i<rboxes.size(); i++){
+    for(size_t i=0; i<rboxes.size(); i++){
         if(isAlmostSquare(rboxes[i])){
             CCheckBox cb;
             cb.outerBBox = rboxes[i];
             cb.validOuterBBox = true;
-            for(int j=0; j<holes.size(); j++){
+            for(size_t j=0; j<holes.size(); j++){
                 Rect inter = rboxes[i] & holes[j];
                 if(inter.area() == holes[j].area()){
                     cb.innerBBox = holes[j];
@@ -326,7 +326,7 @@ bool CDetectCheckBoxes::detectCheckBoxes(string &inFileName, vector<CCheckBox> &
             //rectangle(imgMAT, rboxes[i], Scalar(0,0,255), 2);
         }
     }
-    for(int i=0; i<holes.size(); i++){
+    for(size_t i=0; i<holes.size(); i++){
         if(!usedInner[i]){
             CCheckBox cb;
             cb.innerBBox = holes[i];
@@ -338,7 +338,7 @@ bool CDetectCheckBoxes::detectCheckBoxes(string &inFileName, vector<CCheckBox> &
 
     fillRatioFilter(imgMATbin, cBoxes);
 
-    for(int i=0; i<cBoxes.size(); i++){
+    for(size_t i=0; i<cBoxes.size(); i++){
         CCheckBox cb = cBoxes[i];
         if(cb.innerFillRatio < 0.1){
             cBoxes[i].isFilled = false;
